+++ packages/vibekit/src/services/config-manager.ts
