+++ packages/vibekit/src/services/docker-client.ts
